api/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
requests==2.31.0
pandas==2.2.0
jsonschema==4.21.1
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==42.0.5
pydantic==2.6.1
pyyaml==6.0.1
